+++ src/aks-preview/azext_aks_preview/custom.py
@@ -3109,13 +3109,19 @@ def aks_mesh_enable_egress_gateway(
         client,
         resource_group_name,
         name,
+        istio_egressgateway_name,
+        istio_egressgateway_namespace,
+        gateway_configuration_name,
 ):
     return _aks_mesh_update(
         cmd,
         client,
         resource_group_name,
         name,
-        enable_egress_gateway=True)
+        enable_egress_gateway=True,
+        istio_egressgateway_name=istio_egressgateway_name,
+        istio_egressgateway_namespace=istio_egressgateway_namespace,
+        gateway_configuration_name=gateway_configuration_name)
 
 
 def aks_mesh_disable_egress_gateway(
@@ -3123,12 +3129,16 @@ def aks_mesh_disable_egress_gateway(
         client,
         resource_group_name,
         name,
+        istio_egressgateway_name,
+        istio_egressgateway_namespace,
 ):
     return _aks_mesh_update(
         cmd,
         client,
         resource_group_name,
         name,
+        istio_egressgateway_name=istio_egressgateway_name,
+        istio_egressgateway_namespace=istio_egressgateway_namespace,
         disable_egress_gateway=True)
 
 
@@ -3255,6 +3265,9 @@ def _aks_mesh_update(
         ingress_gateway_type=None,
         enable_egress_gateway=None,
         disable_egress_gateway=None,
+        istio_egressgateway_name=None,
+        istio_egressgateway_namespace=None,
+        gateway_configuration_name=None,
         revision=None,
         yes=False,
         mesh_upgrade_command=None,